Current progress on telling Julia to read survey csv
The idea is to ask what VARIABLES the survey has. Then go through each
question in the survey, and you can say WHICH variables that question
measures. For independent variables (eg Objective) you should be able to
specify that's the question.

The idea is to then take that spec of "these questions measure these
variables" and then just _spit out a DF of observations_