Finish specifying what's dependent, independent
Well, need to test the unhappy path still! But this should allow someone
to specify their variables, and then which are dependent or independent,
with quite wide error tolerance.

I know there are tradeoffs for this - for a lot of variables, you've
gotta scroll back up to check which is dependent and which isn't.
Perhaps I'll do a traditional "name"... "independent? y/N"... thing, and
then this interface at the end