Bit of a refactor, start on columns-tagging
Functions for asking for indices were refactored out, as well as just
asking for variables and doing things with them.

Variables are now mutable cos that makes life a LOT easier in this
stage; I don't have to juggle the index AND value AND array of each
variable. If I need to make the struct immutable again, I'll likely
make a mutable struct that matches it just to allow for the editing
on the fly that I'm doing