Some absolute voodoo
The readAnnotatedData function now returns one of the two dataframes we
want, namely the following columns:
ID Objective Tension ...

It does so by considering each separate question as an individual
variable (I guess two individual unique variables count as one), and
using DataFrames' `stack` function to turn it long; each row containing
each measurement of those variable combos in the whole dataframe. Then I
go through each row and extract the objective and aspect from each, and
well, there we have it.

A note: If I try to broadcast in the @astable part, everything dies, so
I've had to do it the old-fashioned way? We'll see if I can fix that up.

It's also a little brute-force, as it runs that objective-aspect regex
on EACH ROW in that quite-long dataframe.