...Wow. That seems to actually work.
So I already had all the pieces I had to make this generic and a lot
less hacky: as long as that function untangleHeaderText returns a named
tuple, you're good to go, as $AsTable wants a named tuple to dictate
what values to put in the named columns. Excellent. Way less painful to
look at, a lot more generic.